backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
requests>=2.31.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0
drf-spectacular>=0.27.0
bleach>=6.0.0
pytz>=2024.1
hypothesis>=6.0.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
factory-boy>=3.3.0
faker>=19.0.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
factory-boy>=3.3.0
